Python 3.9 support -- IMAP's open takes a timeout argument
For reasons I don't understand, getmail overrides open and does a
monkeypatch instead of using the default python IMAP/SSL open
function.  In any case, the IMAP4_SSL_EXTENDED class's open
did not support the timeout argument.  I added the timeout argument,
but note the monkeypatch version of open doesn't support timeout,
so the timeout argument is discarded.

It allows mail to flow again, but I hardly think it is the correct solution.