Fix for #129
Adds a new option "pipe_stdout" to the [destination] settings block. Settings it to "True" will pipe it to stdout of getmail, "False" will drop stdout from external MDA (like it was before). Default is set to "True" now as this is how fetchmail does.
TODO: Docs needs to be updated, too.